requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
apscheduler==3.10.4
httpx==0.25.1
python-jose==3.3.0
